mv gossip/Service.applyNewState() into separate evm_sync.go
